backend/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Image processing
Pillow==10.1.0
opencv-python-headless==4.8.1.78
numpy==1.26.2

# OCR
pytesseract==0.3.10

# HTTP client for Mistral API
httpx==0.26.0

# Async support
anyio==4.1.0

# Session storage
redis==5.0.1

# Environment variables
python-dotenv==1.0.0

# Utilities
uuid==1.30
